feat: реализованы методы интерфейса CompilationService в классе CompilationServiceImpl; добавлено новое dto UpdateCompilationRequest fix: исправлены ошибки с типом тела запроса эндпоинта Patch /{compilationId}

+++ main-svc/src/main/java/ru/practicum/mainsvc/compilation/controller/CompilationPublicController.java
@@ -26,7 +26,7 @@ public class CompilationPublicController {
     }
 
     @GetMapping("/{compilationId}")
-    public CompilationDto getById(@PathVariable @Positive int compilationId) {
+    public CompilationDto getById(@PathVariable @Positive Long compilationId) {
         return service.getById(compilationId);
     }
 }
